+++ core/initializer/parallelism/test_executor_parallelizer.go
@@ -31,13 +31,12 @@ type ParallelTestOutput struct {
 
 // ================= Parallel executor ============================
 type TestExecutorParallelizer struct {
-	executionId             uuid.UUID
-	dockerClient            *client.Client
-	testControllerImageName string
-	testControllerLogLevel  string
-	testServiceImageName    string
-	testControllerEnvVars   map[string]string
-	parallelism             uint
+	executionId                 uuid.UUID
+	dockerClient                *client.Client
+	testControllerImageName     string
+	testControllerLogLevel      string
+	customTestControllerEnvVars map[string]string
+	parallelism                 uint
 	additionalTestTimeoutBuffer time.Duration
 }
 
@@ -48,8 +47,9 @@ Args:
 	executionId: The UUID uniquely identifying this execution of the tests
 	dockerClient: The handle to manipulating the Docker environment
 	testControllerImageName: The name of the Docker image that will be used to run the test controller
-	testServiceImageName: The name of the Docker image of the version of the service being tested
-	testControllerEnvVars: A custom user-defined map from <env variable name> -> <env variable value> that will be set for test controller
+	testControllerLogLevel: A string, meaningful to the test controller, that represents the user's desired log level
+	customTestControllerEnvVars: A custom user-defined map from <env variable name> -> <env variable value> that will be
+		passed via Docker environment variables to the test controller
 	parallelism: The number of tests to run concurrently
 	additionalTestTimeoutBuffer: The amount of additional timeout given to each test for setup, on top of the test-declared timeout
  */
@@ -58,18 +58,16 @@ func NewTestExecutorParallelizer(
 			dockerClient *client.Client,
 			testControllerImageName string,
 			testControllerLogLevel string,
-			testServiceImageName string,
-			testControllerEnvVars map[string]string,
+			customTestControllerEnvVars map[string]string,
 			parallelism uint,
 			additionalTestTimeoutBuffer time.Duration) *TestExecutorParallelizer {
 	return &TestExecutorParallelizer{
-		executionId:             executionId,
-		dockerClient:            dockerClient,
-		testControllerImageName: testControllerImageName,
-		testControllerLogLevel:  testControllerLogLevel,
-		testServiceImageName:    testServiceImageName,
-		testControllerEnvVars:   testControllerEnvVars,
-		parallelism:             parallelism,
+		executionId:                 executionId,
+		dockerClient:                dockerClient,
+		testControllerImageName:     testControllerImageName,
+		testControllerLogLevel:      testControllerLogLevel,
+		customTestControllerEnvVars: customTestControllerEnvVars,
+		parallelism:                 parallelism,
 		additionalTestTimeoutBuffer: additionalTestTimeoutBuffer,
 	}
 }
@@ -151,8 +149,7 @@ func (executor TestExecutorParallelizer) runTestWorkerGoroutine(
 			testParams.SubnetMask,
 			executor.testControllerImageName,
 			executor.testControllerLogLevel,
-			executor.testServiceImageName,
-			executor.testControllerEnvVars,
+			executor.customTestControllerEnvVars,
 			testParams.TestName,
 			testParams.Test)